+++ snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/datastore/request/SnomedConceptCreateRequest.java
@@ -77,6 +77,10 @@ public final class SnomedConceptCreateRequest extends BaseSnomedComponentCreateR
 	SnomedConceptCreateRequest() {
 	}
 	
+	void setSubclassDefinitionStatus(SubclassDefinitionStatus subclassDefinitionStatus) {
+		this.subclassDefinitionStatus = subclassDefinitionStatus;
+	}
+	
 	void setDefinitionStatus(DefinitionStatus definitionStatus) {
 		this.definitionStatus = definitionStatus;
 	}